wip: Federator process
