+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
 build
-submitter.log
 .DS_Store
+federator.log
