+++ federator/.gitignore
@@ -1,0 +1,8 @@
+config.js
+regtest.json
+development.json
+rinkeby.json
+rsktestnet.json
+test.json
+federator.log
+db/*.*
