+++ federator/config.sample.js
@@ -1,0 +1,6 @@
+module.exports = {
+    mainchain: require('./regtest.json'),
+    sidechain: require('./development.json'),
+    runEvery: 1, // In minutes,
+    confirmations: 5
+}
