+++ federator/log-config.json
@@ -1,0 +1,19 @@
+{
+    "appenders": {
+        "file": {
+            "type": "file",
+            "filename": "federator.log",
+            "maxLogSize": 10485760,
+            "backups": 3,
+            "compress": true
+        },
+        "console": {
+            "type": "console"
+        }
+    },
+    "categories": {
+        "default": {
+            "appenders": ["file", "console"],
+            "level": "debug" }
+    }
+}
