+++ federator/src/lib/CustomError.js
@@ -1,0 +1,8 @@
+class CustomError extends Error {
+    constructor(message, err) {
+        super(message);
+        this.stack += '\n Internal ' + err.stack;
+    }
+}
+
+module.exports = CustomError;
