+++ federator/test/testHelper.js
@@ -1,0 +1,42 @@
+
+advanceTimeAndBlock = async (web3, time) => {
+    await advanceTime(web3,time);
+    await advanceBlock(web3);
+
+    return Promise.resolve(web3.eth.getBlock('latest'));
+}
+
+advanceTime = (web3, time) => {
+    return new Promise((resolve, reject) => {
+        web3.currentProvider.send({
+            jsonrpc: "2.0",
+            method: "evm_increaseTime",
+            params: [time],
+            id: new Date().getTime()
+        }, (err, result) => {
+            if (err) { return reject(err); }
+            return resolve(result);
+        });
+    });
+}
+
+advanceBlock = (web3) => {
+    return new Promise((resolve, reject) => {
+        web3.currentProvider.send({
+            jsonrpc: "2.0",
+            method: "evm_mine",
+            id: new Date().getTime()
+        }, (err, result) => {
+            if (err) { return reject(err); }
+            const newBlockHash = web3.eth.getBlock('latest').hash;
+
+            return resolve(newBlockHash)
+        });
+    });
+}
+
+module.exports = {
+    advanceTime,
+    advanceBlock,
+    advanceTimeAndBlock
+}
