--- submitter/.gitignore
@@ -1,9 +1,0 @@
-config.js
-regtest.json
-development.json
-rinkeby.json
-rinkeby-fork.json
-rsktestnet.json
-test.json
-submitter.log
-mmrDB.json
