--- submitter/log-config.json
@@ -1,19 +1,0 @@
-{
-    "appenders": {
-        "file": {
-            "type": "file",
-            "filename": "submitter.log",
-            "maxLogSize": 10485760, 
-            "backups": 3, 
-            "compress": true
-        },
-        "console": {
-            "type": "console"
-        }
-    },
-    "categories": {
-        "default": {
-            "appenders": ["file", "console"],
-            "level": "debug" }
-    }
-}
