--- submitter/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "submitter",
-  "private": "true",
-  "version": "1.0.0",
-  "description": "Decentralized Bridge Submitter",
-  "dependencies": {
-    "ethereumjs-tx": "^1.3.7",
-    "ethereumjs-util": "^6.1.0",
-    "log4js": "^5.0.0",
-    "web3": "^1.2.1"
-  },
-  "devDependencies": {
-    "chai": "^4.2.0",
-    "mocha": "^6.2.0"
-  },
-  "scripts": {
-    "start": "node ./src/main.js",
-    "test": "mocha --recursive \"test/**/*.test.js\" ",
-    "mmrsync": "node --max-old-space-size=2048 ./src/mmrSync.js",
-    "integrationTest": "node integrationTest.js",
-    "sendTestToken": "node transferTest.js",
-    "clean": "rm -rf ./db/*"
-  },
-  "keywords": [
-    "rsk",
-    "decentralized",
-    "bridge",
-    "ethereum",
-    "submitter"
-  ],
-  "author": "IOVLabs"
-}
