--- submitter/src/abis/MMRProver.json
@@ -1,328 +1,0 @@
-[
-  {
-    "constant": true,
-    "inputs": [],
-    "name": "blockRecorder",
-    "outputs": [
-      {
-        "name": "",
-        "type": "address"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [],
-    "name": "initialBlock",
-    "outputs": [
-      {
-        "name": "",
-        "type": "uint256"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "",
-        "type": "bytes32"
-      }
-    ],
-    "name": "proofs",
-    "outputs": [
-      {
-        "name": "blockNumber",
-        "type": "uint256"
-      },
-      {
-        "name": "blockHash",
-        "type": "bytes32"
-      },
-      {
-        "name": "mmrRoot",
-        "type": "bytes32"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [],
-    "name": "owner",
-    "outputs": [
-      {
-        "name": "",
-        "type": "address"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "inputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "constructor"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "recorder",
-        "type": "address"
-      }
-    ],
-    "name": "setBlockRecorder",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "initial",
-        "type": "uint256"
-      }
-    ],
-    "name": "setInitialBlock",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "blockNumber",
-        "type": "uint256"
-      },
-      {
-        "name": "blockHash",
-        "type": "bytes32"
-      },
-      {
-        "name": "mmrRoot",
-        "type": "bytes32"
-      }
-    ],
-    "name": "getProofId",
-    "outputs": [
-      {
-        "name": "",
-        "type": "bytes32"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "pure",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "blockNumber",
-        "type": "uint256"
-      },
-      {
-        "name": "blockHash",
-        "type": "bytes32"
-      },
-      {
-        "name": "mmrRoot",
-        "type": "bytes32"
-      }
-    ],
-    "name": "initProcessProof",
-    "outputs": [],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": false,
-    "inputs": [
-      {
-        "name": "blockNumber",
-        "type": "uint256"
-      },
-      {
-        "name": "blockHash",
-        "type": "bytes32"
-      },
-      {
-        "name": "mmrRoot",
-        "type": "bytes32"
-      },
-      {
-        "name": "otherBlockNumber",
-        "type": "uint256"
-      },
-      {
-        "name": "initial",
-        "type": "bytes"
-      },
-      {
-        "name": "prefixes",
-        "type": "bytes[]"
-      },
-      {
-        "name": "suffixes",
-        "type": "bytes[]"
-      }
-    ],
-    "name": "processBlockProof",
-    "outputs": [
-      {
-        "name": "",
-        "type": "bool"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "nonpayable",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "blockNumber",
-        "type": "uint256"
-      },
-      {
-        "name": "blockHash",
-        "type": "bytes32"
-      },
-      {
-        "name": "mmrRoot",
-        "type": "bytes32"
-      }
-    ],
-    "name": "isProved",
-    "outputs": [
-      {
-        "name": "",
-        "type": "bool"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "blockNumber",
-        "type": "uint256"
-      },
-      {
-        "name": "blockHash",
-        "type": "bytes32"
-      },
-      {
-        "name": "mmrRoot",
-        "type": "bytes32"
-      }
-    ],
-    "name": "getProofStatus",
-    "outputs": [
-      {
-        "name": "blocksToProve",
-        "type": "uint256[]"
-      },
-      {
-        "name": "proved",
-        "type": "bool[]"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "finalmmr",
-        "type": "bytes32"
-      },
-      {
-        "name": "initial",
-        "type": "bytes"
-      },
-      {
-        "name": "prefixes",
-        "type": "bytes[]"
-      },
-      {
-        "name": "suffixes",
-        "type": "bytes[]"
-      }
-    ],
-    "name": "mmrIsValid",
-    "outputs": [
-      {
-        "name": "",
-        "type": "bool"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "pure",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "blockHash",
-        "type": "bytes32"
-      },
-      {
-        "name": "blockNumber",
-        "type": "uint256"
-      }
-    ],
-    "name": "getBlocksToProve",
-    "outputs": [
-      {
-        "name": "blocksToProve",
-        "type": "uint256[]"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "view",
-    "type": "function"
-  },
-  {
-    "constant": true,
-    "inputs": [
-      {
-        "name": "x",
-        "type": "uint256"
-      }
-    ],
-    "name": "log_2",
-    "outputs": [
-      {
-        "name": "y",
-        "type": "uint256"
-      }
-    ],
-    "payable": false,
-    "stateMutability": "pure",
-    "type": "function"
-  }
-]
