--- submitter/src/lib/CustomError.js
@@ -1,8 +1,0 @@
-class CustomError extends Error {
-    constructor(message, err) {
-        super(message);
-        this.stack += '\n Internal ' + err.stack;
-    }
-}
-
-module.exports = CustomError;
